+++ src/sync/poller.h
@@ -107,7 +107,7 @@ public:
     int UnregisterTimer(int handle) noexcept;
     ffrt_timer_query_t GetTimerStatus(int handle) noexcept;
 
-    uint8_t GetPollCount() noexcept;
+    uint64_t GetPollCount() noexcept;
 
     uint64_t GetTaskWaitTime(CPUEUTask* task) noexcept;
 
@@ -149,7 +149,7 @@ private:
     bool IsTimerReady() noexcept;
 
     int m_epFd;
-    std::atomic<uint8_t> pollerCount_ = 0;
+    std::atomic<uint64_t> pollerCount_ = 0;
     int timerHandle_ = -1;
     std::atomic<EpollStatus> flag_ = EpollStatus::WAKE;
     struct WakeDataWithCb m_wakeData;
